requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
fire==0.5.0
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
multidict==6.0.4
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
termcolor==2.2.0
urllib3==1.26.15
yarl==1.8.2
